+++ user-api/src/main/java/org/ecommerce/userapi/external/controller/UserController.java
@@ -1,12 +1,12 @@
-package org.ecommerce.userapi.controller;
+package org.ecommerce.userapi.external.controller;
 
 import org.ecommerce.common.vo.Response;
 import org.ecommerce.userapi.dto.AccountDto;
 import org.ecommerce.userapi.dto.AddressDto;
 import org.ecommerce.userapi.dto.UserDto;
+import org.ecommerce.userapi.external.service.UserService;
 import org.ecommerce.userapi.security.AuthDetails;
 import org.ecommerce.userapi.security.custom.CurrentUser;
-import org.ecommerce.userapi.service.UserService;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
