+++ user-api/src/main/java/org/ecommerce/userapi/external/service/SellerService.java
@@ -1,4 +1,4 @@
-package org.ecommerce.userapi.service;
+package org.ecommerce.userapi.external.service;
 
 import java.util.Set;
