+++ user-api/src/test/java/org/ecommerce/userapi/controller/UserControllerTest.java
@@ -20,9 +20,9 @@ import org.ecommerce.userapi.entity.Address;
 import org.ecommerce.userapi.entity.Users;
 import org.ecommerce.userapi.entity.UsersAccount;
 import org.ecommerce.userapi.entity.enumerated.Gender;
+import org.ecommerce.userapi.external.service.UserService;
 import org.ecommerce.userapi.repository.UserRepository;
 import org.ecommerce.userapi.security.AuthDetails;
-import org.ecommerce.userapi.service.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
